requirements.txt
# Core MCP library (fast & beginner-friendly wrapper around official MCP Python SDK)
fastmcp>=0.3.0

# HTTP client for calling external APIs (Open-Meteo, NewsAPI, TomTom, Wikipedia, etc.)
httpx>=0.27.0

# Environment variable management (for API keys)
python-dotenv>=1.0.1

# Structured config & validation (optional but very useful)
pydantic>=2.8.0
pydantic-settings>=2.3.0

# For clean type hints & better dev experience
typing-extensions>=4.12.0

# Optional: better logging / debugging during development
rich>=13.7.0

# ────────────────────────────────────────────────
# Optional extras (uncomment if you need them)
# ────────────────────────────────────────────────

# If you want to add unit tests later
# pytest>=8.0.0
# pytest-asyncio>=0.23.0

# If you switch to HTTP transport and want a simple server
# uvicorn>=0.30.0
# fastapi>=0.112.0

# If you add any image processing or charts later
# pillow>=10.0.0